gcf_2/requirements.txt
google-cloud-pubsub==1.5.0
google-cloud-datastore==1.12.0
google-cloud-vision==1.0.0